requirements.txt
annotated-types==0.7.0
anyio==4.12.1
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.38.1
azure-data-tables==12.7.0
azure-functions==1.24.0
azure-search-documents==11.6.0
certifi==2026.1.4
charset-normalizer==3.4.4
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isodate==0.7.2
jiter==0.13.0
MarkupSafe==3.0.3
msrest==0.7.1
multidict==6.7.1
oauthlib==3.3.1
openai==2.20.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
yarl==1.22.0
